requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
greenlet==3.0.3
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pillow==10.2.0
playwright==1.41.2
pyee==11.0.1
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.4.0
soupsieve==2.5
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.9.0
urllib3==2.2.1